+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2022-09-28
+### Added
+- handling of custom audiences
+- added test cases for custom audience
+- preview cookie handling and variation choices
+- bumping to 0.4.0 to keep up with the sst-sdk-node package
+
 ## [0.2.0] - 2022-06-23
 ### Added
 - handle optin cookie
@@ -35,7 +42,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - use PSR-3 for logging
   - use PSR-17,PSR-18 for configuration download
 
-[Unreleased]: https://github.com/SymplifyConversion/sst-sdk-php/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/SymplifyConversion/sst-sdk-php/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/SymplifyConversion/sst-sdk-php/releases/tag/v0.4.0
 [0.2.0]: https://github.com/SymplifyConversion/sst-sdk-php/releases/tag/v0.2.0
 [0.1.1]: https://github.com/SymplifyConversion/sst-sdk-php/releases/tag/v0.1.1
 [0.1.0]: https://github.com/SymplifyConversion/sst-sdk-php/releases/tag/v0.1.0